Subo cosas

+++ Java/ProgramacionJava/src/main/java/JavaOrientadoObjetos/Boletin4/Herencia2/Ejercicio2/Astro.java
@@ -27,7 +27,7 @@ public class Astro {
         this.nombre = nombre;
     }
 
-    public int getMasa() {
+    public long getMasa() {
         return masa;
     }
